Cambios en proceso boleteria para que reciba los ninios por el pipe

+++ ProcesoBoleteria/src/Boleteria.cpp
@@ -1,6 +1,6 @@
 #include "Boleteria.h"
 
-Boleteria::Boleteria(unsigned double precioBoleto, unsigned cantidadBoletos) :
+Boleteria::Boleteria(double precioBoleto, unsigned cantidadBoletos) :
 		caja(), precioBoleto(precioBoleto), cantidadBoletos(cantidadBoletos) {
 }
