+++ ProcesoBoleteria/src/Boleteria.h
@@ -1,16 +1,18 @@
 #ifndef BOLETERIA_H_
 #define BOLETERIA_H_
+#include "Ninio.h"
+#include "Caja.h"
 
 class Boleteria {
 public:
-	Boleteria(unsigned double precioBoleto, unsigned cantidadBoletos);
+	Boleteria(double precioBoleto, unsigned cantidadBoletos);
 	~Boleteria();
 
 	void atenderNinio(const Ninio& ninio);
 
 private:
 	Caja caja;
-	unsigned double precioBoleto;
+	double precioBoleto;
 	unsigned cantidadBoletos;
 };
