+++ ProcesoBoleteria/src/Caja.cpp
@@ -1,25 +1,24 @@
 #include "Caja.h"
 #include <string>
 
-Caja::Caja(unsigned double saldoInicial) :
+Caja::Caja(double saldoInicial) :
 		saldo(saldoInicial) {
 }
 
 Caja::~Caja() {
 }
 
-void Caja::insertarDinero(unsigned double dinero) {
+void Caja::insertarDinero(double dinero) {
 	saldo += dinero;
 }
-void Caja::retirarDinero(unsigned double dinero) {
+void Caja::retirarDinero(double dinero) {
 	if (saldo >= dinero) {
 		saldo -= dinero;
-	}
-	else {
+	} else {
 		throw std::string("Saldo Insuficiente");
 	}
 }
 
-unsigned double Caja::consultarSaldo() const {
+double Caja::consultarSaldo() const {
 	return saldo;
 }
